implement stepper and canopen zaxis hardware modules
